% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topK.R
\name{check_topK}
\alias{check_topK}
\title{Check Top K Predictors}
\usage{
check_topK(X, y, K)
}
\arguments{
\item{X}{A matrix or data frame containing the predictors.}

\item{y}{A vector containing the response variable.}

\item{K}{A positive integer specifying the number of top informative predictors to select; must be greater than the number of predictors. (Optional)}
}
\value{
A list with the updated X matrix and y vector.
}
\description{
This function checks the input of K value if it follow the rules and give different results due to conditions.
}
\details{
This function checks if the input of K is NULL or not.
\cr If K is NULL, the function returns the original X and y without further processing.
\enumerate{ If K is not NULL, the function checks the following conditions before proceeding:
\item P > N: The number of predictors (P) must be less than the number of observations (N).
\item K is numeric: K must be a numeric value.
\item K is greater than zero: K must be greater than zero.
\item K does not exceed the number of predictors: K must not exceed the total number of predictors in X.
}
If all conditions are met, the function proceeds with model regression using the `returnTopK` function(For more information, see \code{\link{returnTopK}}.) to select the top K predictors.
}
\examples{
library(simpleEnsembleGroup7)

# Example usage
X <- iris[, -5]  # predictors
y <- iris[, 5]   # response
result <- check_topK(X, y, 3)
print(result)

}
