% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missing_values.R
\name{handle_missing_values}
\alias{handle_missing_values}
\title{Handle Missing Values}
\usage{
handle_missing_values(X, y)
}
\arguments{
\item{X}{A data frame or matrix containing the predictor variables.}

\item{y}{A vector containing the response variable (Binary or continuous).}
}
\value{
The function a list containing the predictor variables (X) and the response variable (y), after handling missing values.
}
\description{
The handle_missing_values function checks for missing values in the predictor variables (X) and the response variable (y).
Depending on the user's choice, it offers methods to handle missing values such as NA omit removal, mean imputation, median imputation, mode imputation, or retaining NA values.
}
\details{
Missing values are a common issue in datasets and can affect the performance of models if not handled properly.
The `handle_missing_values` function provides several options for handling missing values, including:
\enumerate{
\item  Removing rows with missing values (`na.omit`)
\item  Imputing missing values with the mean
\item  Imputing missing values with the median
\item  Imputing missing values with the mode
\item  Keeping NA values in the dataset for the next step calculation
}
\cr
\cr
For more details about handling missing value, we recommend mice package if you can use.
}
\section{How to Use}{

1. Load the package using `library(my_package)`.
\cr
\cr
2. Prepare your dataset `my_data`, ensuring it contains both predictors and the response variable.
\cr Or you already have the X (predictors) and y (response) variable.
\cr
3. Using `handle_missing_values(X, y)` with your dataset to handle missing values.
\cr
\cr
4. Choose an option to handle missing values:
\enumerate{
  \item   - Enter 1 to remove rows with missing values.
  \item   - Enter 2 to impute missing values with the mean.
  \item   - Enter 3 to impute missing values with the median.
  \item   - Enter 4 to impute missing values with the mode.
  \item   - Enter 5 to keep NA values in the dataset.
}
}

\section{Conclusion}{

Handling missing values is crucial for data analysis and modeling.
By using the `handle_missing_values` function, you can choose the most suitable method for your dataset.
}

\examples{
# Load the package
library(simpleEnsembleGroup7)
X <- matrix(c(1, 2, NA, 4, 5, NA, 2, 3, 4, 5), ncol = 2)
y <- c(1, NA, 3, 4, 5)
data_handled <- handle_missing_values(X, y)
View(data_handled$X)


}
