% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnTopK.R
\name{returnTopK}
\alias{returnTopK}
\title{Top K Informative Predictors and Response}
\usage{
returnTopK(X, y, K)
}
\arguments{
\item{X}{A dataframe containing the predictors. Each column should represent one predictor.}

\item{y}{A vector containing the response variable, which can be either continuous or categorical.}

\item{k}{An integer specifying the number of top informative predictors to return.}
}
\value{
A list of class 'list' containing the following components:
  \itemize{
    \item \code{X}: A dataframe containing the top K informative predictors.
    \item \code{y}: The response variable as provided in the input.
  }
}
\description{
This function evaluates the informativeness of predictors in relation to a given response variable
using appropriate statistical tests based on their data types. It returns the top K informative
predictors along with the response variable for further analysis or modeling.
}
\details{
The function applies different statistical tests depending on the data type of the predictors and the response:
  \itemize{
    \item For continuous predictors and a continuous response, it uses the Pearson Correlation Coefficient.
    \item For categorical predictors and a categorical response, it applies the Chi-square Test of Independence.
    \item For categorical predictors and a continuous response, it uses the Kruskal-Wallis Test.
  }
Note: The function handles NA values by omitting rows with NA in any of the predictors or the response during the analysis.
}
\examples{
data(mtcars)
X <- mtcars[, -which(names(mtcars) == "mpg")]  # Predictor dataset excluding the response
y <- mtcars$mpg  # Response variable
result <- returnTopK(X, y, k = 3)
print(result$X)  # Print the dataframe of top K predictors
print(result$y)  # Print the response variable

}
