% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_model.R
\name{svm_model}
\alias{svm_model}
\title{Support Vector Machine (SVM) Model}
\usage{
svm_model(X, y)
}
\arguments{
\item{X}{Input features (data frame or matrix).}

\item{y}{Response variable (numeric or factor).}
}
\value{
Trained SVM model.
}
\description{
This function trains a support vector machine (SVM) model on the given dataset.
}
\examples{
# Load example dataset
data(iris)
X <- iris[, -which(names(iris) == "Species")]
y <- iris$Species

# Train an SVM model
svm_model <- svm_model(X, y)
svm_model

}
\references{
This function uses the 'svm' function from the 'e1071' package.
}
