% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnTopK.R
\name{var_type}
\alias{var_type}
\title{Determine the Type of a Variable}
\usage{
var_type(x)
}
\arguments{
\item{x}{The variable to type checK.}
}
\value{
A character string describing the type of x.
}
\description{
Internal helper function to determine the data type of a variable.
Used internally by various functions within the pacKage.
}
\keyword{internal}
